+++ internal/repository/transaction.go
@@ -21,7 +21,7 @@ func NewTransactionPostgres(db *sqlx.DB) *TransactionPostgres {
 	}
 }
 
-func (r *TransactionPostgres) GetReceivedTransactions(ctx context.Context, userID int) ([]entity.TransactionDetail, error) {
+func (r *TransactionPostgres) GetReceivedTransactions(ctx context.Context, userID int64) ([]entity.TransactionDetail, error) {
 	var received []entity.TransactionDetail
 	query := `
 		SELECT u.username AS from_user, t.amount
@@ -32,7 +32,7 @@ func (r *TransactionPostgres) GetReceivedTransactions(ctx context.Context, userI
 	return received, r.getter.DefaultTrOrDB(ctx, r.db).SelectContext(ctx, &received, query, userID)
 }
 
-func (r *TransactionPostgres) GetSentTransactions(ctx context.Context, userID int) ([]entity.TransactionDetail, error) {
+func (r *TransactionPostgres) GetSentTransactions(ctx context.Context, userID int64) ([]entity.TransactionDetail, error) {
 	var sent []entity.TransactionDetail
 	query := `
 		SELECT u.username AS to_user, t.amount
@@ -43,7 +43,7 @@ func (r *TransactionPostgres) GetSentTransactions(ctx context.Context, userID in
 	return sent, r.getter.DefaultTrOrDB(ctx, r.db).SelectContext(ctx, &sent, query, userID)
 }
 
-func (r *TransactionPostgres) InsertTransaction(ctx context.Context, fromUserID, toUserID, amount int) error {
+func (r *TransactionPostgres) InsertTransaction(ctx context.Context, fromUserID, toUserID, amount int64) error {
 	query := `INSERT INTO transactions (from_user, to_user, amount) VALUES ($1, $2, $3)`
 	_, err := r.getter.DefaultTrOrDB(ctx, r.db).ExecContext(ctx, query, fromUserID, toUserID, amount)
