+++ internal/repository/user.go
@@ -21,8 +21,8 @@ func NewUserPostgres(db *sqlx.DB) *UserPostgres {
 	}
 }
 
-func (r *UserPostgres) CreateUser(ctx context.Context, user entity.User) (int, error) {
-	var id int
+func (r *UserPostgres) CreateUser(ctx context.Context, user entity.User) (int64, error) {
+	var id int64
 	query := `INSERT INTO users (username, password_hash, coins) VALUES ($1, $2, $3) RETURNING id`
 	row := r.getter.DefaultTrOrDB(ctx, r.db).QueryRowContext(ctx, query, user.Username, user.Password, user.Coins)
 	if err := row.Scan(&id); err != nil {
@@ -39,14 +39,14 @@ func (r *UserPostgres) GetUser(ctx context.Context, username string) (entity.Use
 	return user, r.getter.DefaultTrOrDB(ctx, r.db).GetContext(ctx, &user, query, username)
 }
 
-func (r *UserPostgres) GetUserBalance(ctx context.Context, userID int) (int, error) {
-	var balance int
+func (r *UserPostgres) GetUserBalance(ctx context.Context, userID int64) (int64, error) {
+	var balance int64
 	query := `SELECT coins FROM users WHERE id = $1`
 
 	return balance, r.getter.DefaultTrOrDB(ctx, r.db).GetContext(ctx, &balance, query, userID)
 }
 
-func (r *UserPostgres) UpdateCoins(ctx context.Context, userID, amount int) error {
+func (r *UserPostgres) UpdateCoins(ctx context.Context, userID, amount int64) error {
 	query := `UPDATE users SET coins = coins + $1 WHERE id = $2 AND coins >= $1`
 
 	res, err := r.getter.DefaultTrOrDB(ctx, r.db).ExecContext(ctx, query, amount, userID)
